fix: Исправил составление отчета (#633)
* fix: Исправил составление отчета

Связано: #623

* fix: Исправил редис хост для тестов

Связано: #623

* fix: Исравил конфиг для редиса

Связано: #623
